Refactor: 기존 app.tsx를 loginloading.tsx로 변경 및 app.tsx에 setScreenHeight 함수 설정
