+++ src/App.tsx
@@ -1,18 +1,17 @@
-import { OAUTH_PROVIDER } from '@pages/LoginPage/consts';
-import handleOAuth2Redirect from './pages/LoginPage/OAuth';
-import { useNavigate } from 'react-router';
+import { useEffect } from 'react';
+import { Router } from 'Router';
 function App() {
-    const navigate = useNavigate();
-    const url = new URL(window.location.href);
-    const isNaver = url.searchParams.get('state')?.toString();
-    const isGoogle = url.toString().search('google');
-    if (isNaver === 'nid/me') {
-        handleOAuth2Redirect({ from: OAUTH_PROVIDER.NAVER, navigate });
-    } else if (isGoogle !== -1) {
-        handleOAuth2Redirect({ from: OAUTH_PROVIDER.GOOGLE, navigate });
-    } else {
-        handleOAuth2Redirect({ from: OAUTH_PROVIDER.KAKAO, navigate });
-    }
-    return <div>로그인 중..</div>;
+    const setScreenHeight = () => {
+        const vh = window.innerHeight * 0.01;
+        document.documentElement.style.setProperty('--vh', `${vh}px`);
+    };
+    useEffect(() => {
+        setScreenHeight();
+        window.addEventListener('resize', setScreenHeight);
+        return () => {
+            window.removeEventListener('resize', setScreenHeight);
+        };
+    }, []);
+    return <Router />;
 }
 export default App;
